requirements.txt
prettyprinter == 0.18.*
libPyshell == 0.4.*
pytest == 8.0.*
lark[interegular] == 1.1.*
pydot == 1.2.*
